Add files via upload
This dataset will be the standings of the teams in the last 10 years (excluding teams that weren't available for all 10 years). There are 20 columns, 2 columns will depict of one season. One stat will be the wins which in a regular season will be out of 82 games. The other stat will be the percentage won of all the games (out of 100).